+++ authentication/lib/refinery/authentication.rb
@@ -1,8 +1,10 @@
 require 'refinerycms-core'
+require 'action_mailer'
 require 'devise'
 require 'friendly_id'
 
 module Refinery
+  autoload :AuthenticationGenerator, 'generators/refinery/authentication/authentication_generator'
   autoload :AuthenticatedSystem, 'refinery/authenticated_system'
 
   class << self