+++ authentication/spec/controllers/refinery/admin/users_controller_spec.rb
@@ -1,7 +1,7 @@
 require "spec_helper"
 
 describe Refinery::Admin::UsersController do
-  login_refinery_superuser
+  refinery_login_with :refinery_superuser
 
   shared_examples_for "new, create, update, edit and update actions" do
     it "loads roles" do
@@ -51,7 +51,7 @@ describe Refinery::Admin::UsersController do
 
   describe "#edit" do
     it "renders the edit template" do
-      get :edit, :id => "1"
+      get :edit, :id => logged_in_user.id
       response.should be_success
       response.should render_template("refinery/admin/users/edit")
     end
@@ -60,10 +60,10 @@ describe Refinery::Admin::UsersController do
   end
 
   describe "#update" do
+    let(:additional_user) { FactoryGirl.create :refinery_user }
     it "updates a user" do
-      user = FactoryGirl.create(:refinery_user)
-      Refinery::User.should_receive(:find).at_least(1).times{ user }
-      put "update", :id => user.id.to_s, :user => {}
+      Refinery::User.should_receive(:find).at_least(1).times{ additional_user }
+      put "update", :id => additional_user.id.to_s, :user => {}
       response.should be_redirect
     end