+++ core/lib/generators/refinery/engine/templates/lib/refinery/plural_name/engine.rb
@@ -6,13 +6,10 @@ module Refinery
 
       engine_name :refinery_<%= extension_plural_name %>
 
-      initializer "register refinerycms_<%= plural_name %> plugin" do |app|
+      initializer "register refinerycms_<%= plural_name %> plugin" do
         Refinery::Plugin.register do |plugin|
           plugin.name = "<%= plural_name %>"
-          plugin.url = {
-            :controller => 'refinery/<%= namespacing.underscore %>/admin/<%= plural_name %>',
-            :action => 'index'
-          }
+          plugin.url = proc { Refinery::Core::Engine.routes.url_helpers.<%= namespacing.underscore %>_admin_<%= plural_name %>_path }
           plugin.pathname = root
           plugin.activity = {
             :class_name => :'refinery/<%= namespacing.underscore %>/<%= singular_name %>'<% if (title = attributes.detect { |a| a.type.to_s == "string" }).present? and title.name != 'title' %>,