+++ core/lib/generators/refinery/form/templates/lib/refinery/plural_name.rb
@@ -1,3 +1,4 @@
+require 'refinerycms-settings'
 require 'refinerycms-core'
 
 module Refinery