+++ core/spec/support/refinery.rb
@@ -5,10 +5,15 @@ RSpec.configure do |config|
   config.include Refinery::Testing::ControllerMacros::Methods, :type => :controller
   config.extend Refinery::Testing::RequestMacros::Authentication, :type => :request
   config.include Refinery::Testing::UrlHelper
+  config.include Warden::Test::Helpers
 
   # set some config values so that image and resource factories don't fail to create
   config.before(:each) do
     Refinery::Images.max_image_size = 5242880
     Refinery::Resources.max_file_size = 52428800
+  end
+
+  config.after do
+    Warden.test_reset!
   end
 end