+++ images/config/routes.rb
@@ -1,4 +1,4 @@
-Refinery::Core::Engine.routes.draw do
+Refinery::Core::Engine.routes.prepend do
   match '/system/images/*dragonfly', :to => Dragonfly[:refinery_images]
 
   namespace :admin, :path => 'refinery' do