+++ resources/lib/refinery/resources/engine.rb
@@ -8,7 +8,7 @@ module Refinery
 
       config.autoload_paths += %W( #{config.root}/lib )
 
-      initializer 'setup-refinery-resources-with-dragonfly', :before => :load_config_initializers do |app|
+      initializer 'setup-refinery-resources-with-dragonfly', :before => :load_config_initializers do
         ::Refinery::Resources::Dragonfly.setup!
       end
 
@@ -23,10 +23,8 @@ module Refinery
           plugin.name = 'refinery_files'
           plugin.menu_match = /refinery\/(refinery_)?(files|resources)$/
           plugin.version = %q{2.0.0}
-          plugin.activity = {
-            :class_name => :'refinery/resource'
-          }
-          plugin.url = { :controller => '/refinery/admin/resources' }
+          plugin.activity = { :class_name => :'refinery/resource' }
+          plugin.url = proc { Refinery::Core::Engine.routes.url_helpers.admin_resources_path }
         end
       end