+++ templates/refinery/installer.rb
@@ -1,5 +1,15 @@
 require 'rbconfig'
-VERSION_BAND = '2.0.0'
+VERSION_BAND = '2.0.2'
+
+# We want to ensure that you have an ExecJS runtime available!
+begin
+  run 'bundle install'
+  require 'execjs'
+  ::ExecJS::Runtimes.autodetect
+rescue
+  gsub_file 'Gemfile', "# gem 'therubyracer'", "gem 'therubyracer'"
+end
+
 append_file 'Gemfile' do
 "
 
@@ -7,7 +17,7 @@ append_file 'Gemfile' do
 gem 'refinerycms', '~> #{VERSION_BAND}'
 
 # Specify additional Refinery CMS Extensions here (all optional):
-gem 'refinerycms-i18n',   '~> #{VERSION_BAND}'
+gem 'refinerycms-i18n', '~> #{VERSION_BAND}'
 #  gem 'refinerycms-blog', '~> #{VERSION_BAND}'
 #  gem 'refinerycms-inquiries', '~> #{VERSION_BAND}'
 #  gem 'refinerycms-search', '~> #{VERSION_BAND}'