+++ testing/lib/refinery/testing/request_macros/authentication.rb
@@ -2,29 +2,27 @@ module Refinery
   module Testing
     module RequestMacros
       module Authentication
-        def login_with(factory)
+        def refinery_login_with(factory)
           let!(:logged_in_user) { Factory.create(factory) }
 
           before do
-            visit refinery.new_refinery_user_session_path
-
-            fill_in "Login", :with => logged_in_user.username
-            fill_in "Password", :with => logged_in_user.password
-
-            click_on "Sign in"
+            login_as logged_in_user, :scope => :refinery_user
           end
         end
 
         def login_refinery_user
-          login_with(:refinery_user)
+          Refinery.deprecate(:login_refinery_user, :when => '2.2', :replacement => 'refinery_login_with :refinery_user')
+          refinery_login_with(:refinery_user)
         end
 
         def login_refinery_superuser
-          login_with(:refinery_superuser)
+          Refinery.deprecate(:login_refinery_superuser, :when => '2.2', :replacement => 'refinery_login_with :refinery_superuser')
+          refinery_login_with(:refinery_superuser)
         end
 
         def login_refinery_translator
-          login_with(:refinery_translator)
+          Refinery.deprecate(:login_refinery_translator, :when => '2.2', :replacement => 'refinery_login_with :refinery_translator')
+          refinery_login_with(:refinery_translator)
         end
       end
     end