+++ changelog.md
@@ -1,3 +1,38 @@
+## 2.1.0 [unreleased]
+
+* [See full list](https://github.com/resolve/refinerycms/compare/2-0-stable...master)
+
+## 2.0.3 [unreleased]
+* Fixed missing authentication initializer. [Uģis Ozols](https://github.com/ugisozols)
+* Fixed Heroku and sqlite3 related errors. [Philip Arndt](https://github.com/parndt)
+* Replaced label_tag with label in pages advanced options form. [Johannes Edelstam](https://github.com/jede)
+* Added missing refinerycms-settings require in generated refinery form extension. [Philip Arndt](https://github.com/parndt)
+* Added JS runtime check in templates. [Philip Arndt](https://github.com/parndt) & [Josef Šimánek](https://github.com/simi)
+* Fixed user role assignment issue. [Uģis Ozols](https://github.com/ugisozols)
+* Added image type whitelisting configuration option. [Rob Yurkowski](https://github.com/robyurkowski)
+* Removed global hash of menu instances. [Pete Higgins](https://github.com/phiggins)
+* Fixed save & continue issue. [Philip Arndt](https://github.com/parndt)
+* Fixed issue with Heroku option for CMS generator. [Philip Arndt](https://github.com/parndt)
+* Fixed config.assets.initialize_on_precompile setting to false. [Philip Arndt](https://github.com/parndt)
+
+
+## 2.0.2 [15 March 2012]
+
+* Removed dependencies from refinerycms-testing that were just opinions and not necessary. [Pete Higgins](https://github.com/phiggins)
+* Fixed missing `Refinery::PagePart` positions in seeds. [Mark Stuart](https://github.com/markstuart)
+* Fixed issue with Rakefile template that gets generated into extensions. [Uģis Ozols](https://github.com/ugisozols)
+* Fixed issue where new page parts could not be added to a page. [Uģis Ozols](https://github.com/ugisozols)
+* Added missing initializer for the Authentication extension. [Uģis Ozols](https://github.com/ugisozols)
+* [See full list](https://github.com/resolve/refinerycms/compare/2.0.1...2.0.2)
+
+## 2.0.1 [6 March 2012]
+
+* Updated `plugin.url` code to support Rails 3.2.2. [Philip Arndt](https://github.com/parndt)
+* Added guard-spork '0.5.2' dependency to refinerycms-testing. [Joe Sak](https://github.com/joemsak)
+* Added support for '.' in usernames. [Philip Arndt](https://github.com/parndt)
+* Now includes application.js by default. [Nick Romanowski](https://github.com/nodabs)
+* [See full list](https://github.com/resolve/refinerycms/compare/2.0.0...2.0.1)
+
 ## 2.0.0 [29 February 2012]
 
 * Remove jquery_include_tags helper in favor of using jquery from jquery-rails gem. [Uģis Ozols](https://github.com/ugisozols)